“New Reaction”: problem with the display of molecules to use in the new reaction. Can’t see any.
This problem can occur in Protocol mode or Administration mode.  It happens in the Subcellular Components tab and can happen in the Reactions or Reaction Complex expanders.
The problem is fixed for the Reaction Complex section.  A reaction complex must be selected.  If not, then the "New Reaction" expander will now be hidden.  

Still looking at the "Reactions" section.

git-tfs-id: [http://tfs.bumc.bu.edu:8080/tfs/BU-MICRO-MSI]$/Daphne/Daphne-skg;C5571